refactor: RoomNew 페이지에서 사용되는 폼 로직 분리
