--- src/RoomNew/constants/form.ts
@@ -1,53 +1,0 @@
-import z from 'zod';
-import {
-  FORM_MESSAGE,
-  ROOM_TYPES,
-  TIME_RANGE,
-  ROUTINE_NAME,
-  ROOM_NAME,
-  USER_COUNT,
-  PASSWORD
-} from '@/RoomForm/constants/literals';
-
-export const formSchema = z.object({
-  type: z.enum(ROOM_TYPES),
-  certifyTime: z.number(),
-  routines: z.array(
-    z.object({
-      value: z
-        .string()
-        .trim()
-        .min(ROUTINE_NAME.min, FORM_MESSAGE.ROUTINE_NAME)
-        .max(ROUTINE_NAME.max, FORM_MESSAGE.ROUTINE_NAME)
-    })
-  ),
-  title: z
-    .string()
-    .trim()
-    .min(ROOM_NAME.min, FORM_MESSAGE.ROOM_NAME)
-    .max(ROOM_NAME.max, FORM_MESSAGE.ROOM_NAME),
-  userCount: z
-    .number()
-    .gte(USER_COUNT.min, FORM_MESSAGE.USER_COUNT)
-    .lte(USER_COUNT.max, FORM_MESSAGE.USER_COUNT),
-  password: z.literal('').or(
-    z
-      .string()
-      .min(PASSWORD.min, FORM_MESSAGE.PASSWORD)
-      .max(PASSWORD.max, FORM_MESSAGE.PASSWORD)
-      .refine((v) => /^\d*$/.test(v), {
-        message: FORM_MESSAGE.ONLY_NUMERIC_PASSWORD
-      })
-  )
-});
-
-export const defaultValues: Inputs = {
-  type: 'MORNING',
-  certifyTime: TIME_RANGE['MORNING'][0],
-  routines: [{ value: '' }],
-  userCount: 5,
-  title: '',
-  password: ''
-};
-
-export type Inputs = z.infer<typeof formSchema>;
