+++ src/RoomNew/index.ts
@@ -3,3 +3,5 @@ export { default as PasswordStep } from './steps/PasswordStep';
 export { default as RoutineStep } from './steps/RoutineStep';
 export { default as SummaryStep } from './steps/SummaryStep';
 export { default as TimeStep } from './steps/TimeStep';
+export { default as useFormManagement } from './hooks/useFormManagement';
+export { default as Navbar } from './components/Navbar';
