+++ src/pages/RoomNewPage.tsx
@@ -1,22 +1,14 @@
-import { useMutation } from '@tanstack/react-query';
-import {
-  SubmitErrorHandler,
-  SubmitHandler,
-  useForm,
-  FormProvider
-} from 'react-hook-form';
-import { zodResolver } from '@hookform/resolvers/zod';
-import roomAPI from '@/core/api/functions/roomAPI';
-import { Inputs, defaultValues, formSchema } from '@/RoomNew/constants/form';
+import { FormProvider } from 'react-hook-form';
 import { useFunnel, Funnel } from '@/shared/Funnel';
-import Navbar from '@/RoomNew/components/Navbar';
 import { Header } from '@/shared/Header';
 import {
   BirdStep,
   PasswordStep,
   RoutineStep,
   SummaryStep,
-  TimeStep
+  TimeStep,
+  useFormManagement,
+  Navbar
 } from '@/RoomNew';
 
 export const steps = [
@@ -29,77 +21,13 @@ export const steps = [
 
 const RoomNewPage = () => {
   const funnel = useFunnel(steps);
-
-  const form = useForm<Inputs>({
-    defaultValues,
-    mode: 'onBlur',
-    resolver: zodResolver(formSchema)
-  });
-
-  const { mutate: postRoom, isPending } = useMutation({
-    mutationFn: roomAPI.postRoom
-  });
-
-  const onSubmit: SubmitHandler<Inputs> = (data) => {
-    postRoom(
-      {
-        title: data.title,
-        password: data.password,
-        type: data.type,
-        routine: data.routines.map((r) => r.value),
-        certifyTime: data.certifyTime % 24,
-        maxUserCount: data.userCount
-      },
-      {
-        onSuccess: (data) => {
-          // TODO: 성공 Toast 메시지를 보여줘야 해요.
-          // TODO: toast({ message: '방이 생성되었습니다.', type: 'success' });
-          console.log(data.message);
-
-          // TODO: 방 상세 페이지로 redirect 해야 해요.
-          console.log('TODO: 방 상세 페이지로 redirect');
-        },
-        onError: (error) => {
-          // TODO: 에러 Toast 메시지를 보여줘야 해요.
-          // TODO: toast({ message: '서버에서 날아온 에러 메시지.', type: 'error' });
-          console.log(error.response?.data?.message);
-
-          if (error.response?.data?.validation) {
-            const { setError } = form;
-
-            const {
-              title,
-              password,
-              type,
-              routine,
-              certifyTime,
-              maxUserCount
-            } = error.response.data.validation;
-
-            setError('title', { message: title });
-            setError('password', { message: password });
-            setError('type', { message: type });
-            setError('routines', { message: routine });
-            setError('certifyTime', { message: certifyTime });
-            setError('userCount', { message: maxUserCount });
-          }
-
-          if (error.response?.status === 401) {
-            // TODO: 로그인 페이지로 redirect 해야 해요.
-            console.log('TODO: 로그인 페이지로 redirect');
-          }
-        }
-      }
-    );
-  };
-
-  const onError: SubmitErrorHandler<Inputs> = (errors) => console.error(errors);
+  const { form, mutation, handleSubmit } = useFormManagement();
 
   return (
     <FormProvider {...form}>
       <form
         className="flex h-full flex-col"
-        onSubmit={form.handleSubmit(onSubmit, onError)}
+        onSubmit={handleSubmit}
       >
         <Header
           className="bg-light-main"
@@ -127,7 +55,7 @@ const RoomNewPage = () => {
         </main>
         <Navbar
           {...funnel}
-          isPending={isPending}
+          isPending={mutation.isPending}
         />
       </form>
     </FormProvider>
